services/posts/service/server/requirements.txt
grpcio==1.71.0
grpcio-tools==1.71.0
protobuf>5
googleapis-common-protos==1.69.2